refactor(settings): split DisplaySettingsManager's live half into an excluded companion
performSetup, the connected-display capture, the observer wiring, the
spacing apply with its relaunch wave, the app-modal relaunch
confirmation, and the system-spacing seed move verbatim to
DisplaySettingsManager+Live.swift. Their substance needs a running
AppState, real NSScreen/WindowServer display state, the on-disk
NSStatusItemSpacing global domain, or an NSAlert.

The measured half keeps persistence, lookup, mutation, URI handling, and
every decision rule, including shouldSkipSpacingApply, which the moved
observer consults. Same private-to-internal widening rationale as the
ProfileManager split, noted at the declaration site.
